models/requirements.txt
python-dotenv
langchain
langchain-google-genai
langchain-huggingface
langchain-community
langchain-core
langchain-text-splitters
langgraph
pydantic
chromadb
langdetect